MA-8805 - On evk6sl, the actual recorded duration is much less.
In evk6sl, sometimes overrun when record.
To recover from overrun, the pipe reader discard 1/16 of the max
frames, which is 32 in the case.
But unfortunatelly, the write unit is 48 frames. If next time the
pipe writer is first schduled, overrun occurs again.
To fix it, when overrun, discard reasonable frames.

Change-Id: Ia1c4283ae523a7035afc1866c2054cea042f1c99
Signed-off-by: fang hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1433
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>